% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscheme2.R
\name{color_scheme_2}
\alias{color_scheme_2}
\title{colorscheme2}
\usage{
color_scheme_2(seq, scheme_df)
}
\arguments{
\item{seq}{aligned data frame}

\item{scheme_df}{color scheme data frame}
}
\value{
color matched with sequence
}
\description{
colorscheme2
}
\examples{
color_scheme_2(sampleseq,chemi_aa_coldf)
}
\author{
huina huang
}
