% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documention.r
\docType{data}
\name{sampleseq}
\alias{sampleseq}
\title{sample sequences(amino acids) aligned}
\format{A data frame with 9 rows and 20 variables}
\usage{
data(sampleseq)
}
\description{
A dataset containing the sample sequence of 9 observations of 20
variables. The variables are as follows:
}
\details{
\itemize{
   \item{\code{V1}}{a factor with levels \code{-} \code{i} \code{q}}
   \item{\code{V2}}{a factor with levels \code{-} \code{a} \code{t} \code{v}}
   \item{\code{V3}}{a factor with levels \code{-} \code{c} \code{l} \code{y}}
   \item{\code{V4}}{a factor with levels \code{-} \code{n} \code{r}}
   \item{\code{V5}}{a factor with levels \code{-} \code{a} \code{p} \code{y}}
   \item{\code{V6}}{a factor with levels \code{-} \code{d} \code{r}}
   \item{\code{V7}}{a factor with levels \code{-} \code{f} \code{h} \code{w}}
   \item{\code{V8}}{a factor with levels \code{-} \code{g} \code{t}}
   \item{\code{V9}}{a factor with levels \code{-} \code{i} \code{l} \code{m} \code{q} \code{v}}
   \item{\code{V10}}{a factor with levels \code{a} \code{d} \code{e} \code{p}}
   \item{\code{V11}}{a factor with levels \code{i} \code{n} \code{q}}
   \item{\code{V12}}{a factor with levels \code{d} \code{g} \code{p}}
   \item{\code{V13}}{a factor with levels \code{f} \code{l} \code{r} \code{v} \code{w} \code{y}}
   \item{\code{V14}}{a factor with levels \code{d} \code{e} \code{h} \code{i} \code{v}}
   \item{\code{V15}}{a factor with levels \code{-} \code{e} \code{h} \code{r} \code{t}}
   \item{\code{V16}}{a factor with levels \code{y}}
   \item{\code{V17}}{a factor with levels \code{m} \code{p} \code{s} \code{t}}
   \item{\code{V18}}{a factor with levels \code{a} \code{d} \code{e} \code{q}}
   \item{\code{V19}}{a factor with levels \code{a} \code{e} \code{t}}
   \item{\code{V20}}{a factor with levels \code{d} \code{e}}
}
}
\keyword{datasets}
