% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documention.r
\docType{data}
\name{sampleseqcolor_chemistry}
\alias{sampleseqcolor_chemistry}
\title{color of sampleseq according to chemi_nucle_coldf}
\format{A data frame with 9 rows and 20 variables}
\usage{
data(sampleseqcolor_chemistry)
}
\description{
A dataset containing the test sequence of 9 observations of 20
variables. The variables are as follows:
}
\details{
\itemize{
   \item{\code{V1}}{a character vector}
   \item{\code{V2}}{a character vector}
   \item{\code{V3}}{a character vector}
   \item{\code{V4}}{a character vector}
   \item{\code{V5}}{a character vector}
   \item{\code{V6}}{a character vector}
   \item{\code{V7}}{a character vector}
   \item{\code{V8}}{a character vector}
   \item{\code{V9}}{a character vector}
   \item{\code{V10}}{a character vector}
   \item{\code{V11}}{a character vector}
   \item{\code{V12}}{a character vector}
   \item{\code{V13}}{a character vector}
   \item{\code{V14}}{a character vector}
   \item{\code{V15}}{a character vector}
   \item{\code{V16}}{a character vector}
   \item{\code{V17}}{a character vector}
   \item{\code{V18}}{a character vector}
   \item{\code{V19}}{a character vector}
   \item{\code{V20}}{a character vector}
}
}
\keyword{datasets}
