% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documention.r
\docType{data}
\name{Taylor_aa_coldf}
\alias{Taylor_aa_coldf}
\title{color scheme of sequence which are amino acids aligned according to Taylor}
\format{A data frame with 20 rows and 2 variables}
\usage{
data(Taylor_aa_coldf)
}
\description{
A dataset containing the color scheme of 20 observations of 2
variables. The variables are as follows:
}
\details{
\itemize{
   \item{\code{amino_acid}}{a character vector}
   \item{\code{color}}{a character vector}
}
}
\keyword{datasets}
